functions/requirements.txt
firebase-functions
firebase-admin
google-generativeai
google-cloud-firestore
weasyprint
Flask
Flask-Cors
requests
beautifulsoup4